+++ src/main/java/br/com/itall/model/dao/cad/UsuarioDAO.java
@@ -12,7 +12,7 @@ import javax.naming.NamingException;
 
 import br.com.itall.model.dao.util.ConnectDAO;
 import br.com.itall.model.dao.util.SqlPreparedStatementUpdate;
-import br.com.itall.model.dto.UsuarioDTO;
+import br.com.itall.model.dto.cad.UsuarioDTO;
 import br.com.itall.model.entity.cad.UsuarioModel;
 import br.com.itall.tool.Texto;
 
@@ -116,7 +116,7 @@ public class UsuarioDAO extends ConnectDAO {
 	/**
 	 * Lista todos os usuários em ordem alfabética
 	 * 
-	 * @return List&lt;Usuario&gt;
+	 * @return List&lt;UsuarioDTO&gt;
 	 * @throws SQLException Trata erros de execução do SQL
 	 * @throws NamingException Trata erros de sintaxe do SQL
 	 */
@@ -151,7 +151,7 @@ public class UsuarioDAO extends ConnectDAO {
 	/**
 	 * Verifica se um determinado e-mail existe na tabela de usuários
 	 * 
-	 * @param email (String) Conta de e-ail
+	 * @param email (String) Conta de e-mail
 	 * @return boolean
 	 * @throws SQLException Trata erros de execução SQL.
 	 */
